mapping 등 수정
