+++ .settings/org.eclipse.core.resources.prefs
@@ -1,7 +1,6 @@
 eclipse.preferences.version=1
 encoding//src/main/java=UTF-8
 encoding//src/main/resources=UTF-8
-encoding//src/main/webapp/WEB-INF/exerciser/registerForm.css=UTF-8
 encoding//src/main/webapp/WEB-INF/exerciser/registerForm.jsp=UTF-8
 encoding//src/main/webapp/index.jsp=UTF-8
 encoding//src/test/java=UTF-8
