+++ src/main/webapp/index.jsp
@@ -1,5 +1,5 @@
 <html>
 <body>
-<h2>Hello World! 호정 하이</h2>
+<h2>Hello World!</h2>
 </body>
 </html>
